Accept fractional points, re-check the head before deferring to the webhook, and recompute repair plans
- record_autograder_head_metadata declared p_points as integer, but `points` in pawtograder.yml is a
  plain number and the validator accepts fractions, so a template scoring 1.5 made PostgREST unable
  to resolve the call at all ("function ... (numeric) does not exist"). This path treats that as
  fatal before publishing grader_repo, so a valid solution repository would stay permanently
  unattached until somebody changed the scoring. The parameter is numeric now, which reproduces the
  pre-existing push-webhook behaviour exactly — that path writes the same value into
  assignments.autograder_points, a bigint, so the fraction rounds on assignment. Verified: 1.5 now
  applies and stores 2, matching a direct write.

- webhookHasReconciled required only that the pointer name this repo and that SOME sha be recorded.
  A targeted repair of a repo DELETED on GitHub after its pointer was set — a shape the repair
  script explicitly supports — satisfied both, so this function fetched the replacement's head and
  config and then discarded them, reporting success while the database kept a config and a sha that
  exist in no repository. It now also requires the recorded sha to BE the fetched head, which moves
  the decision below the head lookup.

- The repair script's apply-time guard only ran when the plan already included the handout step, so
  it was one-directional: a plan built as solution-only stayed that way even after an instructor
  switched the assignment to fork_from_prior_assignment and picked a source, and the solution
  endpoint — which accepts every repository-backed mode — would publish grader_repo while
  template_repo still named the old handout. `functions` is recomputed from the reloaded row now.

- The script also lacked the reconciler's isConfiguredFork exception, so a fork with a named source
  and no pointer yet went to needsReview and was silently skipped without --check-github, even
  though repairing it only inherits an existing repository. planFor gains the matching clause the
  reconciler already had: a NULL pointer needs the handout call whatever the autograder setting.

+++ supabase/migrations/20260909140000_record_autograder_head_metadata.sql
@@ -19,7 +19,15 @@ CREATE OR REPLACE FUNCTION public.record_autograder_head_metadata(
     p_expected_sha text,
     p_new_sha text,
     p_config jsonb,
-    p_points integer,
+    -- numeric, not integer. `points` in pawtograder.yml is typed as a plain number and the
+    -- validator accepts fractions, so calculateTotalAutograderPoints can legitimately return 1.5.
+    -- Declared as integer, PostgREST could not resolve the call at all ("function ... (numeric)
+    -- does not exist") — and this path treats that as fatal before publishing grader_repo, so a
+    -- perfectly good solution repository would stay permanently unattached until somebody changed
+    -- the scoring. The pre-existing push-webhook path writes the same value straight into
+    -- assignments.autograder_points, which is bigint, so the fraction is rounded on assignment;
+    -- numeric here reproduces that behaviour exactly rather than inventing a second one.
+    p_points numeric,
     p_message text,
     p_author text,
     p_ref text,
@@ -110,8 +118,8 @@ BEGIN
 END;
 $$;
 
-REVOKE EXECUTE ON FUNCTION public.record_autograder_head_metadata(bigint, text, text, jsonb, integer, text, text, text, text, boolean) FROM PUBLIC;
-GRANT EXECUTE ON FUNCTION public.record_autograder_head_metadata(bigint, text, text, jsonb, integer, text, text, text, text, boolean) TO service_role;
+REVOKE EXECUTE ON FUNCTION public.record_autograder_head_metadata(bigint, text, text, jsonb, numeric, text, text, text, text, boolean) FROM PUBLIC;
+GRANT EXECUTE ON FUNCTION public.record_autograder_head_metadata(bigint, text, text, jsonb, numeric, text, text, text, text, boolean) TO service_role;
 
 ----------------------------------------------------------------------------------------
 -- publish_grader_repo: attach the solution pointer, or refuse
